Whoo, okay, so all the tests work again
So what had happened was I forgot to make the `generateScanArgs`
methods for ProductAttribute and ProductAttributeValue operate by
reference, so while sqlmock was doing its job, it was doing its job
on the wrong kinds of variables, which is 100% my fault and holy
crap it took me like 5 hours to figure this shit out. Kill me now.