Fix product attribute existence by name check
previously, we just checked if an attribute with the same name
already existed in the database before disallowing attribute creation.
This was the wrong approach, because product attributes are inherently
tied to products. So if product A had attribute X, and I wanted product
B to have the same attribute, the old code would have failed, which
is obviously bad. This code should work, though.