Made all tests execute in parallel
This significantly sped up testing, if you count that testing time
used to take .068 seconds and now takes 0.44 seconds. Most of the
test overhead is running everything in docker anyway, but it's good
to get in the habit because there are likely to be hundreds of tests
once all is said and done.