Fix README generation tool
Previously I had the README generation tool figure out what the
current commit hash was, use it to generate dairycart github links
for that specific commit hash url, and all was fine. But then I had
the problem where it would never re-link existing lines because it
wasn't looking for them. I've decided the best way to go around this
is to just link to master, which you ordinarily shouldn't do. But
the alternative is having integration test additions be a two-commit
process that I can't even promise I'd adhere to, much less that
I could convince others to adhere to.