Add integration tests for product attribute value updates
Also added a warning to the README generation tool that complains
if you miss a spot.

I feel like the attribute value update function should warn you if
you're trying to create a value that already exists for that attribute
(the way the creation function does), but that does seem quite a bit
like handholding, so at the very least I'll put it on the backburner
trello thing so I know about it for later.