Make product list queries have an accurate count
Previously we just returned the length of products returned in the
result, which was always the same as the limit, so like, who cares?

Now we use fancy stackoverflow postgres tricks to get a count of all
variables, though curiously this doesn't work when the offset is set
too high. I've decided that's a tradeoff I'm willing to make. Also
in this change is that I've made it so you can't set a page value to
zero, which I'm glad I thought of because that would have been pretty
infuriating to discover later IMO