A rough shot at a (currently non-functional) tool to check coverage
Go's built-in coverage tool only checks that every line of a function
is called at a given point. I'd like to be able to assert that every
function has a test specifically for it as well. This test only does
so much because I'm on a plane and can't google simple shit in Go
(like how to delete things from a slice, lol). But I do want to get
this at least committed locally so I can move on with my life for
the remainder of this flight.