Copmletely reorganize dairycart
So I think the API, integration tests, etc, all need to be their own
repo with their own histories. This is a step in that process. First,
get every subfolder to have its own vendor, make sure all the scripts
work, etc. Then, create new repos in github for all of these. The
organization is at least a little bit goofy for now, but it'll have
to do while I get things sorted.