Parallelize product integration tests (#4)
* Parallelize product integration tests, remove example data

Previously, we relied on existing state of certain dummy values in
the integration tests. This meant that only certain tests which were
non-mutating could be safely marked as parallely executable. In the
longer run there should be as many of these tests are are necessary.
If most of them have to be run in a particular order it means their
organization is hyper-important and, of more concern, these tests
slow down significantly. This change makes it so that every test has
most of its prerequisites created as a part of its setup, and those
prerequisites are also torn down as part of the test. There may be
a better way to accomplish this goal, but I'm not a testing engineer,
so feel free to issue a PR.