Add bridge table for products and values, create products from options
There's a ton of stuff in this commit, and I don't even think I'm done yet. Basically, I've mostly changed the way we're creating products, by automatically creating variant products when we create new products with option data. This creates a few interesting challenges, though, notably how do you go about handling the use case where a user adds a new option value to an existing entry. The simple answer is to just archive all the existing products, but that can be intensely disruptive and wholly unnecessary. So now we have a bridge table that lets us keep track of what variants already exist, so we can hopefully, in the previously described case, just create the newly relevant products. There are some instances, however, where we will have to archive and regenerate products (for instance, upon deletion of an option)