Return a single form of data when creating products
Previously, I had made it so that if you created a product that had no options, it responded with whatever the old API would have responded with. But if you created a product with options, it would respond with an array. This didn't make sense, and would probably have frustrated the hell out of me if I was an integrator. So I changed it. Also attached original option values to the cartesian product result.